refactor :: 관심분야 토글 타입 변경

+++ core/data/src/main/java/team/retum/data/repository/interests/InterestsRepository.kt
@@ -4,7 +4,7 @@ import team.retum.network.model.response.interests.FetchInterestsRecruitmentsRes
 import team.retum.network.model.response.interests.FetchInterestsResponse
 
 interface InterestsRepository {
-    suspend fun setInterestsToggle(codes: List<Int>)
+    suspend fun setInterestsToggle(codes: List<Long>)
     suspend fun fetchInterests(): FetchInterestsResponse
     suspend fun fetchInterestsSearchRecruitments(): FetchInterestsRecruitmentsResponse
 }
